More User Functions
This commit adds the ability to find a single user based on that user's property: value. This function will only return one single user, so if the value of the property is not unique, no more than one entry will be returened.

MAC OS dependencies have been 'fixed' for the time being. There may be a possible confliction between Linux bcrypt and MAC bcrypt. That will be tested and explored hopefully by the next commit.