findAll and update
This update adds the findAll and update functions to the user controller. findAll simply returnes a list of all users stored in the database (in no particular order (yet)). The update function will update a user's information provided the target user's _id is supplied in the request url and a property: newvalue is supplied in the body.