Post Refrences and 'update'
This update adds support for including the author of a post in the response from the server. The included user object has had the password field excluded on the schema level, meaning that there will never ever be an uncaught password leak.

The user schema has been updated and now the rank property works properly, ensuring that the supplied value only matches a list of acceptable values.