Heavily Polished Schedule Controller
Added more enhancements and bug-fixes to the schedule controller. Schedules, so long as they are not for the current day, can now be submitted for the future. The server will check all submitted schedules to make sure that there aren't any conflictions. Schould a schedule be submitted for the current day, and there is not already one stored in the cache, it will be saved and then loaded for that day.