Authentication Update and Schedule Tweaks
This update brings some slight authentication tweaks and improvements over the lat implementation. With this update, a user's rank now plays a role in what API functions they are allowed to access. If a user does not have the required rank, their request will be rejected.

Schedules have been tweaked to properly trash the scheduleCache when the delete request is sent. Also, schedule posted for the current day will now be recognised and loaded into the scheduleCache after entry into the database!

I'm sure there's more in there, but I lost track...